master/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "master"
version = "0.1.0"
description = "NexGen Master LLM Orchestrator"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "httpx>=0.27",
  "pydantic-settings>=2.2",
  "redis[hiredis]>=5.0",
  "structlog>=24.1",
  "nexgen-shared",
  "pytest-asyncio>=1.3.0",
  "openai>=2.31.0",
  "tiktoken>=0.12.0",
  "streamlit>=1.56.0",
  "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.uv.sources]
# Resolve shared contract package from the mono-repo, not PyPI.
nexgen-shared = { path = "../nexgen_shared" }
